Added GUI for extraction of Data

+++ gui.py
@@ -3,6 +3,7 @@ from tkinter import ttk
 from tkinter.scrolledtext import ScrolledText
 import tkinter.messagebox
 import subprocess
+import os
 
 
 def sendMailCallBack():
@@ -17,27 +18,40 @@ def sendMailCallBack():
         print(isinstance(cont, str))
         """
 
-        f = open("plain", "w+")
-        f.write(cont)
-        fs = open("subject", "w+")
-        fs.write(sub)
-        ft = open("to", "w+")
-        ft.write(to)
-        f.close()
-        fs.close()
-        ft.close()
+        if to == "":
+            var = tk.messagebox.showinfo("Send Mail", "Enter valid to Address")
+            return
 
-        processText = subprocess.Popen(["./processText.sh", "plain"])
-        processText.wait()
+        elif sub == "":
+            var = tk.messagebox.showinfo("Send Mail", "Enter valid Subject")
+            return
 
-        sendMail = subprocess.Popen(["./sendMail.sh", "to", "subject"])
-        sendMail.wait()
+        elif cont == "":
+            var = tk.messagebox.showinfo("Send Mail", "Enter valid Content")
+            return
 
-        var = tk.messagebox.showinfo("Send Mail", "Succesfully Sent Mail!!")
+        else:
+            f = open("plain", "w+")
+            f.write(cont)
+            fs = open("subject", "w+")
+            fs.write(sub)
+            ft = open("to", "w+")
+            ft.write(to)
+            f.close()
+            fs.close()
+            ft.close()
 
-        a1.delete(0, tk.END)
-        b1.delete(0, tk.END)
-        c1.delete("1.0", tk.END)
+            processText = subprocess.Popen(["./processText.sh", "plain"])
+            processText.wait()
+
+            sendMail = subprocess.Popen(["./sendMail.sh", "to", "subject"])
+            sendMail.wait()
+
+            var = tk.messagebox.showinfo("Send Mail", "Succesfully Sent Mail!!")
+
+            a1.delete(0, tk.END)
+            b1.delete(0, tk.END)
+            c1.delete("1.0", tk.END)
 
     def closeWindow():
         smWin.destroy()
@@ -65,6 +79,62 @@ def sendMailCallBack():
     ).place(relx=0.55, rely=0.85, anchor=tk.CENTER)
 
 
+def extractDataCallBack():
+    def closeWindow():
+        edWin.destroy()
+
+    def extractFunction():
+        reloc = loc.get()
+
+        if(reloc == ""):
+            var = tk.messagebox.showinfo("Send Mail", "Enter valid Path")
+            return
+
+        else: 
+            dirPath=os.path.dirname(reloc)
+            #print(dirPath)
+
+            logPath = os.path.join(dirPath, "embedlog" + "." + "logAES")
+            #print(logPath)
+
+            extractText = subprocess.Popen(["./extract.sh", logPath, reloc])
+            extractText.wait()
+
+            reProcessText = subprocess.Popen(["./reProcessText.sh"])
+            reProcessText.wait()
+
+            fileP = open("test1", "r")
+            textData = fileP.read()
+
+            content.insert(tk.INSERT, textData)
+
+
+
+
+
+
+    edWin = tk.Toplevel(window)
+    edWin.title("Send Mail")
+    edWin.geometry("600x400")
+    edWin.configure(background="grey")
+
+    a = tk.Label(edWin, text="Location: ").grid(row=0, column=0)
+    loc = tk.Entry(edWin, width="40")
+    loc.grid(row=0, column=1)
+
+    d1 = tk.Button(edWin, text="Extract", highlightbackground="#3E4149", command=extractFunction)
+    d1.grid(row=0, column=2)
+
+    b = tk.Label(edWin, text="Email Content: ").grid(row=1, column=0)
+
+    content = ScrolledText(edWin, width=50, height=15, selectborderwidth=2)
+    content.grid(row=1, column=1)
+
+    btn = ttk.Button(edWin, text="Back", command=closeWindow).place(
+        relx=0.5, rely=0.75, anchor=tk.CENTER
+    )
+
+
 window = tk.Tk()
 window.title("Welcome to Secure Mail Sender")
 window.geometry("400x400")
@@ -78,7 +148,7 @@ def closeWindow():
 btn = ttk.Button(window, text="Send Mail", command=sendMailCallBack).place(
     relx=0.5, rely=0.35, anchor=tk.CENTER
 )
-btn = ttk.Button(window, text="Extract Data").place(
+btn = ttk.Button(window, text="Extract Data", command=extractDataCallBack).place(
     relx=0.5, rely=0.55, anchor=tk.CENTER
 )
 btn = ttk.Button(window, text="Quit", command=closeWindow).place(
